+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 * Support Command Templating and copy to clipboard.
 * Create a progress update window to update the progress while downloading packages and extract man pages.
 * Add command deletion button to each command in the list.
-* Detect user input for CommandBodyBox and CommentBox and create a save button in such events.
+* Detect user input for CommandBodyBox and CommentBox and create a save button in such events.]
+* Add a button beside the command body to copy the command to clipboard.
 
 ### CHANGED
 * Use static images for some buttons